+++ src/frontend/admin/src/components/templates/courses/form/sections/offer/OfferRow.tsx
@@ -15,77 +15,77 @@ import {
   OfferRuleDummy,
 } from "@/services/api/models/OfferRule";
 import { DefaultRow } from "@/components/presentational/list/DefaultRow";
-import { CourseProductRelation } from "@/services/api/models/Relations";
+import { Offer } from "@/services/api/models/Offers";
 import { useModal } from "@/components/presentational/modal/useModal";
 import { CustomModal } from "@/components/presentational/modal/Modal";
-import { OfferRuleForm } from "@/components/templates/courses/form/sections/product-relation/OfferRuleForm";
+import { OfferRuleForm } from "@/components/templates/courses/form/sections/offer/OfferRuleForm";
 import { Maybe } from "@/types/utils";
-import { useCourseProductRelations } from "@/hooks/useCourseProductRelation/useCourseProductRelation";
+import { useOffers } from "@/hooks/useOffer/useOffer";
 import { AlertModal } from "@/components/presentational/modal/AlertModal";
 import { useList } from "@/hooks/useList/useList";
-import { OfferRuleRow } from "@/components/templates/courses/form/sections/product-relation/OfferRuleRow";
+import { OfferRuleRow } from "@/components/templates/courses/form/sections/offer/OfferRuleRow";
 import { CustomLink } from "@/components/presentational/link/CustomLink";
 import { PATH_ADMIN } from "@/utils/routes/path";
 import { MenuPopover } from "@/components/presentational/menu-popover/MenuPopover";
 import { useCopyToClipboard } from "@/hooks/useCopyToClipboard";
 import { commonTranslations } from "@/translations/common/commonTranslations";
-import { CourseProductRelationSource } from "@/components/templates/relations/course-product-relation/CourseProductRelationList";
-import { CourseProductRelationRepository } from "@/services/repositories/course-product-relation/CourseProductRelationRepository";
+import { OfferSource } from "@/components/templates/offers/offer/OfferList";
+import { OfferRepository } from "@/services/repositories/offer/OfferRepository";
 
 const messages = defineMessages({
   mainTitleOfferRule: {
-    id: "components.templates.courses.form.productRelation.row.mainTitleOfferRule",
+    id: "components.templates.courses.form.offer.row.mainTitleOfferRule",
     description: "Title for the offer rule row",
     defaultMessage: "Offer rule {number}",
   },
   subTitleOfferRule: {
-    id: "components.templates.courses.form.productRelation.row.subTitleOfferRule",
+    id: "components.templates.courses.form.offer.row.subTitleOfferRule",
     description: "Sub title for the offer rule row",
     defaultMessage: "{reservedSeats}/{totalSeats} seats",
   },
   addOfferRuleButton: {
-    id: "components.templates.courses.form.productRelation.row.addOfferRuleButton",
+    id: "components.templates.courses.form.offer.row.addOfferRuleButton",
     description: "Add offer rule button label",
     defaultMessage: "Add offer rule",
   },
   offerRuleIsActiveSwitchAriaLabel: {
-    id: "components.templates.courses.form.productRelation.row.offerRuleIsActiveSwitchAriaLabel",
+    id: "components.templates.courses.form.offer.row.offerRuleIsActiveSwitchAriaLabel",
     description: "Aria-label for the offer rule is active switch",
     defaultMessage: "Offer rule is active switch",
   },
   deleteOfferRuleModalTitle: {
-    id: "components.templates.courses.form.productRelation.row.deleteOfferRuleModal",
+    id: "components.templates.courses.form.offer.row.deleteOfferRuleModal",
     description: "Title for the delete offer rule modal",
     defaultMessage: "Delete an offer rule",
   },
   deleteOfferRuleModalContent: {
-    id: "components.templates.courses.form.productRelation.row.deleteOfferRuleModalContent",
+    id: "components.templates.courses.form.offer.row.deleteOfferRuleModalContent",
     description: "Content for the delete offer rule modal",
     defaultMessage: "Are you sure you want to delete this offer rule?",
   },
-  relationDisabledActionsMessage: {
-    id: "components.templates.courses.form.productRelation.row.relationDisabledActionsMessage",
-    description: "Information message for relation disabled actions",
+  offerDisabledActionsMessage: {
+    id: "components.templates.courses.form.offer.row.offerDisabledActionsMessage",
+    description: "Information message for offer disabled actions",
     defaultMessage:
-      "One or more orders have already occurred for this relationship, so you cannot perform this action",
+      "One or more orders have already occurred for this offer, so you cannot perform this action",
   },
   addOfferRuleModalFormTitle: {
-    id: "components.templates.courses.form.productRelation.row.addOfferRuleModalFormTitle",
+    id: "components.templates.courses.form.offer.row.addOfferRuleModalFormTitle",
     defaultMessage: "Add an offer rule",
     description: "Title for the add offer rule modal",
   },
   editOfferRuleModalFormTitle: {
-    id: "components.templates.courses.form.productRelation.row.addOfferRuleModalFormTitle",
+    id: "components.templates.courses.form.offer.row.addOfferRuleModalFormTitle",
     defaultMessage: "Edit an offer rule",
     description: "Title for the edit offer rule modal",
   },
   generateCertificate: {
-    id: "components.templates.courses.form.productRelation.row.generateCertificate",
+    id: "components.templates.courses.form.offer.row.generateCertificate",
     defaultMessage: "Generate certificates",
     description: "Label for the generate certificate action",
   },
   alreadyCertificateGenerationInProgress: {
-    id: "components.templates.courses.form.productRelation.row.alreadyCertificateGenerationInProgress",
+    id: "components.templates.courses.form.offer.row.alreadyCertificateGenerationInProgress",
     defaultMessage: "There is already a certificate generation in progress",
     description:
       "Text when hovering over the action to generate certificates, but a generation is already in progress",
@@ -99,35 +99,28 @@ type EditOfferRuleState = {
 
 type Props = {
   loading?: boolean;
-  relation: CourseProductRelation;
-  onClickEdit: (relation: CourseProductRelation) => void;
-  onClickDelete: (relation: CourseProductRelation) => void;
-  source: CourseProductRelationSource;
+  offer: Offer;
+  onClickEdit: (offer: Offer) => void;
+  onClickDelete: (offer: Offer) => void;
+  source: OfferSource;
   invalidateCourse: () => void;
 };
 
-export function CourseProductRelationRow({
-  relation,
-  onClickEdit,
-  onClickDelete,
-  source,
-}: Props) {
+export function OfferRow({ offer, onClickEdit, onClickDelete, source }: Props) {
   const intl = useIntl();
 
   const jobQuery = useQuery({
-    queryKey: ["course-product-relation-job", relation.id],
+    queryKey: ["offer-job", offer.id],
     staleTime: 0,
     queryFn: async () => {
-      return CourseProductRelationRepository.checkStatutCertificateGenerationProcess(
-        relation.id,
-      );
+      return OfferRepository.checkStatutCertificateGenerationProcess(offer.id);
     },
   });
 
   const copyToClipboard = useCopyToClipboard();
-  const canEdit = relation.can_edit;
+  const canEdit = offer.can_edit;
   const disabledActionsMessage = canEdit
-    ? intl.formatMessage(messages.relationDisabledActionsMessage)
+    ? intl.formatMessage(messages.offerDisabledActionsMessage)
     : undefined;
 
   const [currentOfferRule, setCurrentOfferRule] =
@@ -137,21 +130,16 @@ export function CourseProductRelationRow({
     useList<OfferRuleDummy>([]);
 
   const { items: offerRuleList, ...offerRuleListMethods } = useList<OfferRule>(
-    relation.offer_rules ?? [],
+    offer.offer_rules ?? [],
   );
 
   const offerRuleModal = useModal();
   const deleteOfferRuleModal = useModal();
 
-  const courseProductRelationQuery = useCourseProductRelations(
-    {},
-    { enabled: false },
-  );
+  const offerQuery = useOffers({}, { enabled: false });
 
   const sendGenerateCertificate = async () => {
-    await CourseProductRelationRepository.generateMultipleCertificate(
-      relation.id,
-    );
+    await OfferRepository.generateMultipleCertificate(offer.id);
     await jobQuery.refetch();
   };
 
@@ -160,7 +148,7 @@ export function CourseProductRelationRow({
     offerRule: OfferRule,
     index: number,
   ) => {
-    const { course_product_relation: courseId, ...restPayload } = payload;
+    const { offer: courseId, ...restPayload } = payload;
     const { id: offerRuleId } = offerRule;
     offerRuleListMethods.updateAt(index, {
       ...offerRule,
@@ -169,13 +157,13 @@ export function CourseProductRelationRow({
         (offerRule.nb_available_seats ?? 0) +
         ((payload.nb_seats ?? 0) - (offerRule.nb_seats ?? 0)),
     });
-    courseProductRelationQuery.methods.editOfferRule(
+    offerQuery.methods.editOfferRule(
       {
-        relationId: relation.id,
+        offerId: offer.id,
         offerRuleId,
         payload: {
           ...payload,
-          course_product_relation: relation.id,
+          offer: offer.id,
         },
       },
       {
@@ -193,9 +181,9 @@ export function CourseProductRelationRow({
   const deleteOfferRule = (offerRule: OfferRule, index: number) => {
     const { id: offerRuleId } = offerRule;
     offerRuleListMethods.removeAt(index);
-    courseProductRelationQuery.methods.deleteOfferRule(
+    offerQuery.methods.deleteOfferRule(
       {
-        relationId: relation.id,
+        offerId: offer.id,
         offerRuleId,
       },
       {
@@ -224,9 +212,9 @@ export function CourseProductRelationRow({
     };
 
     dummyListMethods.push(dummy);
-    courseProductRelationQuery.methods.addOfferRule(
+    offerQuery.methods.addOfferRule(
       {
-        relationId: relation.id,
+        offerId: offer.id,
         payload,
       },
       {
@@ -241,37 +229,37 @@ export function CourseProductRelationRow({
   };
 
   useEffect(() => {
-    offerRuleListMethods.set(relation.offer_rules);
-  }, [relation]);
+    offerRuleListMethods.set(offer.offer_rules);
+  }, [offer]);
 
   const getMainTitle = (): ReactNode => {
-    if (source === CourseProductRelationSource.PRODUCT) {
+    if (source === OfferSource.PRODUCT) {
       return (
-        <CustomLink href={PATH_ADMIN.courses.edit(relation.course!.id)}>
-          {relation.course!.title}
+        <CustomLink href={PATH_ADMIN.courses.edit(offer.course!.id)}>
+          {offer.course!.title}
         </CustomLink>
       );
     }
     return (
-      <CustomLink href={PATH_ADMIN.products.edit(relation.product!.id)}>
-        {relation.product!.title}
+      <CustomLink href={PATH_ADMIN.products.edit(offer.product!.id)}>
+        {offer.product!.title}
       </CustomLink>
     );
   };
 
   const getTitle = (): string => {
-    return source === CourseProductRelationSource.COURSE
-      ? relation.product!.title
-      : relation.course!.title;
+    return source === OfferSource.COURSE
+      ? offer.product!.title
+      : offer.course!.title;
   };
 
   return (
     <>
       <DefaultRow
-        loading={courseProductRelationQuery.states.updating}
+        loading={offerQuery.states.updating}
         key={getTitle()}
         mainTitle={getMainTitle()}
-        subTitle={relation.organizations.map((org) => org.title).join(",")}
+        subTitle={offer.organizations.map((org) => org.title).join(",")}
         enableEdit={canEdit}
         enableDelete={canEdit}
         disableDeleteMessage={disabledActionsMessage}
@@ -285,14 +273,14 @@ export function CourseProductRelationRow({
               >
                 <AccessTimeIcon
                   sx={{ ml: 1 }}
-                  data-testid={`already-generate-job-${relation.id}`}
+                  data-testid={`already-generate-job-${offer.id}`}
                   fontSize="medium"
                   color="disabled"
                 />
               </Tooltip>
             )}
             <MenuPopover
-              id={`course-product-relation-actions-${relation.id}`}
+              id={`offer-actions-${offer.id}`}
               menuItems={[
                 {
                   mainLabel: intl.formatMessage(messages.generateCertificate),
@@ -306,15 +294,15 @@ export function CourseProductRelationRow({
                 {
                   mainLabel: intl.formatMessage(commonTranslations.copyUrl),
                   icon: <CopyAllIcon fontSize="small" />,
-                  onClick: () => copyToClipboard(relation.uri!),
+                  onClick: () => copyToClipboard(offer.uri!),
                 },
               ]}
             />
           </>
         }
         disableEditMessage={disabledActionsMessage}
-        onEdit={() => onClickEdit(relation)}
-        onDelete={() => onClickDelete(relation)}
+        onEdit={() => onClickEdit(offer)}
+        onDelete={() => onClickDelete(offer)}
       >
         <Stack gap={2}>
           {offerRuleList.map((offerRule, orderIndex) => {
@@ -336,7 +324,7 @@ export function CourseProductRelationRow({
                     {
                       is_active: isActive,
                       nb_seats: offerRule.nb_seats,
-                      course_product_relation: relation.id,
+                      offer: offer.id,
                       discount_id: offerRule.discount?.id ?? null,
                     },
                     { ...offerRule },
@@ -385,7 +373,7 @@ export function CourseProductRelationRow({
               offerRuleModal.handleClose();
               const payload: DTOOfferRule = {
                 ...values,
-                course_product_relation: relation.id,
+                offer: offer.id,
               };
               if (currentOfferRule) {
                 update(